Module libimagstore::store

Structs

Entry
An Entry of the store

EntryHeader
EntryHeader

FileLockEntry
A struct that allows you to borrow an Entry

Store
The Store itself, through this object one can interact with IMAG's entries

Walk

Enums

StoreObject

Type Definitions

EntryContent
EntryContent type

EntryResult

Result
The Result Type returned by any interaction with the store that could fail