Struct libimagstore::store::Store

pub struct Store { /* fields omitted */ }

The Store itself, through this object one can interact with IMAG's entries

Methods

impl Store

fn new(location: PathBuf, store_config: Option<Value>) -> Result<Store>

Create a new Store object

fn config(&self) -> Option<&Value>

Get the store configuration

fn create<'a, S: IntoStoreId>(&'a self, id: S) -> Result<FileLockEntry<'a>>

Creates the Entry at the given location (inside the entry)

fn retrieve<'a, S: IntoStoreId>(&'a self, id: S) -> Result<FileLockEntry<'a>>

Borrow a given Entry. When the FileLockEntry is either updated or dropped, the new Entry is written to disk

Implicitely creates a entry in the store if there is no entry with the id id. For a non-implicitely-create look at Store::get.

fn get<'a, S: IntoStoreId + Clone>(&'a self,
id: S)
-> Result<Option<FileLockEntry<'a>>>

Get an entry from the store if it exists.

This executes the {pre,post}_retrieve_aspects hooks.

fn retrieve_for_module(&self, mod_name: &str) -> Result<StoreIdIterator>

Iterate over all StoreIds for one module name

fn walk<'a>(&'a self, mod_name: &str) -> Walk

fn update<'a>(&'a self, entry: FileLockEntry<'a>) -> Result<()>

Return the FileLockEntry and write to disk

fn retrieve_copy<S: IntoStoreId>(&self, id: S) -> Result<Entry>

Retrieve a copy of a given entry, this cannot be used to mutate the one on disk

fn delete<S: IntoStoreId>(&self, id: S) -> Result<()>

Delete an entry

fn save_to(&self, entry: &FileLockEntry, new_id: StoreId) -> Result<()>

Save a copy of the Entry in another place Executes the post_move_aspects for the new id

fn save_as(&self, entry: FileLockEntry, new_id: StoreId) -> Result<()>

Save an Entry in another place Removes the original entry Executes the post_move_aspects for the new id

fn move_by_id(&self, old_id: StoreId, new_id: StoreId) -> Result<()>

Move an entry without loading

This function moves an entry from one path to another.

Generally, this function shouldn't be used by library authors, if they "just" want to move something around. A library for moving entries while caring about meta-data and links.

Errors

This function returns an error in certain cases:

If pre-move-hooks error (if they return an error which indicates that the action should be aborted)

If the about-to-be-moved entry is borrowed

If the lock on the internal data structure cannot be aquired

If the new path already exists

If the about-to-be-moved entry does not exist

If the FS-operation failed

If the post-move-hooks error (though the operation has succeeded then).

Warnings

This should be used with great care, as moving an entry from a to b might result in dangling links (see below).

Moving linked entries

If the entry which is moved is linked to another entry, these links get invalid (but we do not detect this here). As links are always two-way-links, so a is not only linked to b, but also the other way round, moving b to c results in the following scenario:

a links to b, which does not exist anymore.

c links to a, which does exist.

So the link is partly dangling, so to say.

fn path(&self) -> &PathBuf

Gets the path where this store is on the disk

fn register_hook(&mut self,
position: HookPosition,
aspect_name: &str,
h: Box<Hook>)
-> Result<()>