Struct libimagstore::storeid::StoreId

pub struct StoreId {
    // some fields omitted
}

The Index into the Store

Methods

impl StoreId

fn new(base: Option<PathBuf>, id: PathBuf) -> Result<StoreId>

fn from_full_path<D>(store_part: &PathBuf, full_path: D) -> Result<StoreId> where D: Deref<Target=Path>

Try to create a StoreId object from a filesystem-absolute path.

Automatically creates a StoreId object which has a base set to store_part if stripping the store_part from the full_path succeeded.

Returns a StoreErrorKind::StoreIdBuildFromFullPathError if stripping failes.

fn new_baseless(id: PathBuf) -> Result<StoreId>

fn without_base(self) -> StoreId

fn with_base(self, base: PathBuf) -> Self

fn into_pathbuf(self) -> Result<PathBuf>

Transform the StoreId object into a PathBuf, error if the base of the StoreId is not specified.

fn exists(&self) -> bool

fn to_str(&self) -> Result<String>

fn components(&self) -> Components

Returns the components of the id part of the StoreId object.

Can be used to check whether a StoreId points to an entry in a specific collection of StoreIds.

fn local(&self) -> &PathBuf

Get the local part of a StoreId object, as in "the part from the store root to the entry".