Struct libimagstore::store::Store

pub struct Store {
    // some fields omitted
}

The Store itself, through this object one can interact with IMAG's entries

Methods

impl Store

fn new(location: PathBuf, store_config: Option<Value>) -> Result<Store>

Create a new Store object

fn config(&self) -> Option<&Value>

Get the store configuration

fn create<'a, S: IntoStoreId>(&'a self, id: S) -> Result<FileLockEntry<'a>>

Creates the Entry at the given location (inside the entry)

fn retrieve<'a, S: IntoStoreId>(&'a self, id: S) -> Result<FileLockEntry<'a>>

Borrow a given Entry. When the FileLockEntry is either updated or dropped, the new Entry is written to disk

Implicitely creates a entry in the store if there is no entry with the id id. For a non-implicitely-create look at Store::get.

fn get<'a, S: IntoStoreId + Clone>(&'a self, id: S) -> Result<Option<FileLockEntry<'a>>>

Get an entry from the store if it exists.

This executes the {pre,post}_retrieve_aspects hooks.

fn retrieve_for_module(&self, mod_name: &str) -> Result<StoreIdIterator>

Iterate over all StoreIds for one module name

fn walk<'a>(&'a self, mod_name: &str) -> Walk

fn update<'a>(&'a self, entry: FileLockEntry<'a>) -> Result<()>

Return the FileLockEntry and write to disk

fn retrieve_copy<S: IntoStoreId>(&self, id: S) -> Result<Entry>

Retrieve a copy of a given entry, this cannot be used to mutate the one on disk

fn delete<S: IntoStoreId>(&self, id: S) -> Result<()>

Delete an entry

fn save_to(&self, entry: &FileLockEntry, new_id: StoreId) -> Result<()>

Save a copy of the Entry in another place Executes the post_move_aspects for the new id

fn save_as(&self, entry: FileLockEntry, new_id: StoreId) -> Result<()>

Save an Entry in another place Removes the original entry Executes the post_move_aspects for the new id

fn move_by_id(&self, old_id: StoreId, new_id: StoreId) -> Result<()>

Move an entry without loading

fn path(&self) -> &PathBuf

Gets the path where this store is on the disk

fn register_hook(&mut self, position: HookPosition, aspect_name: &str, h: Box<Hook>) -> Result<()>